make outputs more verbose to overcome new CASA behavior of not printing out tclean commands

+++ alma/reduction/slurm_subjob_jwbrick.py
@@ -34,42 +34,55 @@ mses = os.getenv('MSES').split() or ['uid___A002_Xf287d3_Xcd1e.ms','uid___A002_X
 logprint(f"Working in {tmpdir} based on files in {workdir}, using MSes ${mses}")
 logprint(f"Variables are: startchan={start}, nchan={nchan}, spw={spw}, mous={mous}, field={field}, fieldname={fieldname}, tmpdir={tmpdir}, fulltmpdir={fulltimpdir}")
 
-splitnames = [msname.replace(".ms", f'_{spw}.split') for msname in mses]
-for msname in mses:
-    split(f'{workdir}/{msname}', msname.replace(".ms", f'_{spw}.split'),
-          spw=str(spw),
-          datacolumn='corrected', field=field)
+if os.path.exists(os.path.join(workdir,
+                               f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.image')):
+    results = glob.glob(f'{workdir}/{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual*')
+    logprint(f"Found completed results {results}.  Finishing without doing any more work.")
+else:
+    splitnames = [msname.replace(".ms", f'_{spw}.split') for msname in mses]
+    for msname in mses:
+        split_kwargs = dict(vis=f'{workdir}/{msname}',
+                            outputvis=msname.replace(".ms", f'_{spw}.split'),
+                            spw=str(spw),
+                            datacolumn='corrected',
+                            field=field)
+        logprint(f"split kwargs: {split_kwargs}")
 
-tclean(vis=splitnames,
-       imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual',
-       field=field,
-       start=start,
-       nchan=nchan,
-       specmode='cube',
-       threshold='0.1mJy',
-       imsize=[6000,7000],
-       cell=['0.03arcsec'],
-       niter=1000,
-       deconvolver='hogbom',
-       phasecenter='J2000 17:46:19.157 -028.35.15.041',
-       gridder='mosaic',
-       weighting='briggs',
-       robust=0.5,
-       pbcor=True,
-       pblimit=0.2,
-       interactive=False)
+        split(**split_kwargs)
+
+    tclean_kwargs = dict(vis=splitnames,
+           imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual',
+           field=field,
+           start=start,
+           nchan=nchan,
+           specmode='cube',
+           threshold='0.1mJy',
+           imsize=[6000,7000],
+           cell=['0.03arcsec'],
+           niter=1000,
+           deconvolver='hogbom',
+           phasecenter='J2000 17:46:19.157 -028.35.15.041',
+           gridder='mosaic',
+           weighting='briggs',
+           robust=0.5,
+           pbcor=True,
+           pblimit=0.2,
+           interactive=False)
+    logprint(f'Tclean kwargs: {tclean_kwargs}')
+
+    tclean(**tclean_kwargs)
 
 
 
-#Create fits datacubes for science targets
-exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.image.pbcor', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.image.pbcor.fits')
-exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.pb', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.pb.fits')
-exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.mask', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.mask.fits')
+    #Create fits datacubes for science targets
+    exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.image.pbcor', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.image.pbcor.fits')
+    exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.pb', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.pb.fits')
+    exportfits(imagename=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.mask', fitsimage=f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual.mask.fits')
 
-for fn in glob.glob(f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual*'):
-    logprint(f"Moving {fn} to {workdir}")
-    shutil.move(fn, workdir)
+    for fn in glob.glob(f'{mous}.{field}_sci.spw{spw}.{start}-{start+nchan}.cube.I.manual*'):
+        logprint(f"Moving {fn} to {workdir}")
+        shutil.move(fn, workdir)
 
-for fn in splitnames:
-    logprint(f"Removing {fn}")
-    shutil.rmtree(fn)
+    for fn in splitnames:
+        logprint(f"Removing {fn}")
+        shutil.rmtree(fn)
